feat: set up restore/create wizard step
