+++ packages/web/src/new/scenes/Main/Auth/Auth.VewModel.ts
@@ -1,17 +1,25 @@
-import { makeObservable, observable } from 'mobx';
+import { action, makeObservable, observable } from 'mobx';
 import { singleton } from 'tsyringe';
 
 import { ViewModel } from 'new/core/viewmodels/ViewModel';
+import type { WizardPayload } from 'new/scenes/Main/Auth/typings';
+import { WizardSteps } from 'new/scenes/Main/Auth/typings';
 
 @singleton()
 export class AuthVewModel extends ViewModel {
-  isAuthenticated = true;
+  step: WizardSteps;
+  static defaultState = {
+    step: WizardSteps.CHOOSE_FLOW,
+  };
 
   constructor() {
     super();
 
+    this.step = AuthVewModel.defaultState.step;
+
     makeObservable(this, {
-      isAuthenticated: observable,
+      step: observable,
+      onWizardChange: action.bound,
     });
   }
 
@@ -24,6 +32,12 @@ export class AuthVewModel extends ViewModel {
   }
 
   protected override setDefaults() {
-    // @TODO why to duplicate
+    this.step = AuthVewModel.defaultState.step;
+  }
+
+  onWizardChange(payload: WizardPayload) {
+    // @TODO
+    this.step = payload.step;
+    // console.log(payload);
   }
 }
