--- packages/web/src/new/scenes/Main/Auth/steps/wizard.tsx
@@ -1,9 +1,0 @@
-import type { FC } from 'react';
-
-export interface Props {}
-
-export const Wizard: FC<Props> = (props) => {
-  return <div></div>;
-};
-
-export default Wizard;
