+++ packages/web/src/new/scenes/Main/Auth/typings.ts
@@ -1,0 +1,13 @@
+import type { ReactElement } from 'react';
+
+export enum WizardSteps {
+  CHOOSE_FLOW = 'CHOOSE_FLOW',
+}
+
+export type WizardPayload = {
+  step: WizardSteps;
+};
+
+export type ViewMap = {
+  [K in WizardSteps]: ReactElement;
+};
